# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClinicalFlow API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: List schedules with filters
      parameters:
        - name: studentId
          in: query
          schema:
            type: string
        - name: ciId
          in: query
          schema:
            type: string
        - name: hospitalId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, active, completed, cancelled]
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Schedule"
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Create a schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
  /schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Get schedule detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Edit schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
    delete:
      operationId: cancelSchedule
      tags: [schedules]
      summary: Cancel schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled

  # ── ATTENDANCE ────────────────────────────────────────────────────────────
...
  /recommendations:
    get:
      operationId: getRecommendations
      tags: [recommendations]
      summary: Get ranked recommended students for a duty or slot
      parameters:
        - name: scheduleId
          in: query
          schema:
            type: string
        - name: slotId
          in: query
          schema:
            type: string
        - name: hospitalId
          in: query
          schema:
            type: string
        - name: departmentId
          in: query
          schema:
            type: string
        - name: caseIds
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Ranked student recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRecommendation"
  /recommendations/explanation:
    get:
      operationId: getRecommendationExplanation
      tags: [recommendations]
      summary: Get AI-generated explanation for a student recommendation score
      parameters:
        - name: studentId
          in: query
          required: true
          schema:
            type: string
        - name: scheduleId
          in: query
          schema:
            type: string
        - name: slotId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationExplanation"

  # ── NOTIFICATIONS ─────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, role, firstName, lastName, isActive, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, scheduler, ci, student]
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        ciProfile:
          $ref: "#/components/schemas/CIProfile"
...
    StudentProfile:
      type: object
      required: [id, userId, studentNumber]
      properties:
        id:
          type: string
        userId:
          type: string
        studentNumber:
          type: string
        yearLevel:
          type: ["integer", "null"]
        section:
          type: ["string", "null"]
        program:
          type: ["string", "null"]
        academicYear:
          type: ["string", "null"]
        totalHoursRequired:
          type: integer
        createdAt:
          type: string
          format: date-time
    CIProfile:
      type: object
      required: [id, userId]
      properties:
        id:
          type: string
        userId:
          type: string
        employeeId:
          type: ["string", "null"]
        specialization:
          type: ["string", "null"]

    # ── STUDENT DETAILS ────────────────────────────────────────────────────
    StudentDetail:
      type: object
      required: [id, firstName, lastName, email, isActive]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        studentProfile:
          $ref: "#/components/schemas/StudentProfile"
        totalHoursCompleted:
          type: number
        totalHoursRequired:
          type: integer
        attendanceRate:
          type: number
        caseCompletionRate:
          type: number
        absenceCount:
          type: integer
        lateCount:
          type: integer
        needsMakeup:
          type: boolean
        isAtRisk:
          type: boolean
...
    Hospital:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: ["string", "null"]
        contactNumber:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        attendanceRadius:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        departments:
          type: array
          items:
            $ref: "#/components/schemas/Department"
...
    Department:
      type: object
      required: [id, hospitalId, name, isActive]
      properties:
        id:
          type: string
        hospitalId:
          type: string
        name:
          type: string
        code:
          type: ["string", "null"]
        isActive:
          type: boolean
...
    Schedule:
      type: object
      required: [id, dutyDate, startTime, endTime, status, gracePeriodMin, createdAt]
      properties:
        id:
          type: string
        title:
          type: ["string", "null"]
        hospitalId:
          type: string
        departmentId:
          type: string
        ciId:
          type: string
        dutyDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        gracePeriodMin:
          type: integer
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        hospital:
          $ref: "#/components/schemas/Hospital"
        department:
          $ref: "#/components/schemas/Department"
        ci:
          $ref: "#/components/schemas/User"
        students:
          type: array
          items:
            $ref: "#/components/schemas/User"
    ScheduleInput:
      type: object
      required: [hospitalId, departmentId, ciId, dutyDate, startTime, endTime, studentIds]
      properties:
        title:
          type: string
        hospitalId:
          type: string
        departmentId:
          type: string
        ciId:
          type: string
        dutyDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        gracePeriodMin:
          type: integer
        notes:
          type: string
        studentIds:
          type: array
          items:
            type: string
    ScheduleUpdate:
      type: object
      properties:
        title:
          type: string
        hospitalId:
          type: string
        departmentId:
          type: string
        ciId:
          type: string
        dutyDate:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        gracePeriodMin:
          type: integer
        notes:
          type: string
        studentIds:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [upcoming, active, completed, cancelled]

    # ── ATTENDANCE ─────────────────────────────────────────────────────────
...
    StudentRecommendation:
      type: object
      required: [studentId, score, reasons]
      properties:
        studentId:
          type: string
        score:
          type: number
        reasons:
          type: array
          items:
            $ref: "#/components/schemas/RecommendationReason"
        student:
          $ref: "#/components/schemas/StudentDetail"
    RecommendationReason:
      type: object
      required: [criterion, weight, applied]
      properties:
        criterion:
          type: string
        weight:
          type: number
        applied:
          type: boolean
        description:
          type: string
    RecommendationExplanation:
      type: object
      required: [studentId, score, explanation]
      properties:
        studentId:
          type: string
        score:
          type: number
        explanation:
          type: string

    # ── NOTIFICATIONS ──────────────────────────────────────────────────────
...